:recycle: refactor: 카드 컴포넌트 통합 및 경로 변경
