+++ src/stories/card.tsx
@@ -1,2 +1,2 @@
-export { default as CardBase } from "@/src/shared/ui/card.base";
-export type { CardProps } from "@/src/shared/ui/card.base";
+export { default as CardBase } from "@/src/shared/ui/card";
+export type { CardProps } from "@/src/shared/ui/card";
